minor: Improved general layout & implemented icon nav-buttons
This is a rather big commit containing overall quality-of-life improvements, as well as many small features which were missing in the previous release of the frontend. Note this is still WIP and therefore not finished! (Things like the welcome text, editor sizing and wave layout still need to be fixed)