fix: Fixed NGINX routing issues and added support for .html extensions
This commit fixes the issue that NGINX would return a 404.html page for static paths that are not '/' (index). This was due to the fact Vue is a single-page application and only exports a single html file called index.html, as such paths such as `/editor` would not be resolved if they are called statically i.e. a page were to be reloaded or a user puts in a full path.